Move CRTAuxDetInitializer declaration to a header file.
Right now, CRTAuxDetInitializer is defined in an unnamed namespace. This commit adds a public interface.
